+++ CHANGES.md
@@ -1,5 +1,14 @@
 # Change log
 
+## Unrelease
+
+- Authentik user provisioning
+- Upgrade notes:
+    - New backend environment vars (see .env-example):
+        - AUTHENTIK_API_URL
+	- AUTHENTIK_TOKEN
+	- AUTHENTIK_GROUP_ID
+
 ## 1.0.0 (2024-07-26)
 
 - Invoices